requirements.txt
pmaw
praw
pytest
mysql-connector
emoji
python-dotenv